+++ example/a.ts
@@ -2,6 +2,8 @@ import { Elysia, t, getSchemaValidator, fileType } from '../src'
 import { z } from 'zod'
 import { post } from '../test/utils'
 
-console.log(Response.json({
-	a: 'a'
-}))
+console.log(
+	Response.json({
+		a: 'a'
+	})
+)