+++ example/a.ts
@@ -1,3 +1,22 @@
 import { Elysia } from '../src'
+import { mapEarlyResponse } from '../src/adapter/web-standard/handler'
 
-new Elysia().get('/', 'Hi').listen(3000)
+const body = {
+	name: 'Shiroko'
+}
+
+const context = {
+	headers: {
+		'x-powered-by': 'Elysia',
+		'coffee-scheme': 'Coffee'
+	},
+	status: 418,
+	cookie: {}
+}
+
+const response = await mapEarlyResponse(
+	new Promise((resolve) => resolve(body)),
+	context
+)
+
+console.log(response?.headers)